+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -119,6 +119,11 @@ New checks
   Finds unintended character output from ``unsigned char`` and ``signed char``
   to an ``ostream``.
 
+- New :doc:`misra-array-declaration
+  <clang-tidy/checks/misra/array-declaration>` check.
+
+  FIXME: Write a short description.
+
 - New :doc:`readability-ambiguous-smartptr-reset-call
   <clang-tidy/checks/readability/ambiguous-smartptr-reset-call>` check.